# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AccountCredentials {
  /**
     * @minLength 3
     * @maxLength 16
     * @pattern ^[A-Za-z0-9_]+$
     */
  username: string;
  /**
     * @minLength 10
     * @maxLength 128
     */
  password: string;
}

export type PlayerPresence = typeof PlayerPresence[keyof typeof PlayerPresence];


export const PlayerPresence = {
  online: 'online',
  offline: 'offline',
  'in-game': 'in-game',
} as const;

export interface Player {
  id: string;
  username: string;
  presence: PlayerPresence;
  characterSkin: string;
  isGuest: boolean;
}

export interface AuthResult {
  user: Player;
}

export interface ActionResult {
  success: boolean;
}

export interface UsernameInput {
  /**
     * @minLength 3
     * @maxLength 16
     */
  username: string;
}

export type FriendPresence = typeof FriendPresence[keyof typeof FriendPresence];


export const FriendPresence = {
  online: 'online',
  offline: 'offline',
  'in-game': 'in-game',
} as const;

export interface Friend {
  userId: string;
  username: string;
  presence: FriendPresence;
}

export interface FriendRequest {
  id: string;
  userId: string;
  username: string;
  createdAt: string;
}

export type PartyInputMode = typeof PartyInputMode[keyof typeof PartyInputMode];


export const PartyInputMode = {
  'coop-squad': 'coop-squad',
  'team-race': 'team-race',
} as const;

export interface PartyInput {
  mode: PartyInputMode;
  isPrivate: boolean;
}

export interface RoomCodeInput {
  /**
     * @minLength 6
     * @maxLength 6
     * @pattern ^[A-Z0-9]{6}$
     */
  roomCode: string;
}

export type PartyMemberPresence = typeof PartyMemberPresence[keyof typeof PartyMemberPresence];


export const PartyMemberPresence = {
  online: 'online',
  offline: 'offline',
  'in-game': 'in-game',
} as const;

export interface PartyMember {
  userId: string;
  username: string;
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSlot: number;
  ready: boolean;
  isLeader: boolean;
  presence: PartyMemberPresence;
}

export type PartyMode = typeof PartyMode[keyof typeof PartyMode];


export const PartyMode = {
  'coop-squad': 'coop-squad',
  'team-race': 'team-race',
} as const;

export type PartyStatus = typeof PartyStatus[keyof typeof PartyStatus];


export const PartyStatus = {
  lobby: 'lobby',
  starting: 'starting',
  'in-game': 'in-game',
} as const;

export interface Party {
  id: string;
  roomCode: string;
  mode: PartyMode;
  isPrivate: boolean;
  status: PartyStatus;
  members: PartyMember[];
}

export interface PartyResult {
  party: Party | null;
}

export interface PartyInvitation {
  id: string;
  partyId: string;
  roomCode: string;
  fromUsername: string;
  createdAt: string;
}

export interface ReadyInput {
  ready: boolean;
}

export interface TeamInput {
  /**
     * @minimum 1
     * @maximum 4
     */
  teamSlot: number;
}

export type ModeInputMode = typeof ModeInputMode[keyof typeof ModeInputMode];


export const ModeInputMode = {
  'coop-squad': 'coop-squad',
  'team-race': 'team-race',
} as const;

export interface ModeInput {
  mode: ModeInputMode;
}

export interface LobbySummary {
  friendsOnline: number;
  incomingFriendRequests: number;
  partyInvitations: number;
}