# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCredentials"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /auth/login:
    post:
      operationId: loginUser
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCredentials"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
...
  /friends/requests:
    get:
      operationId: getFriendRequests
      tags: [friends]
      responses:
        "200":
          description: Incoming requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
    post:
      operationId: sendFriendRequest
      tags: [friends]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsernameInput"
      responses:
        "201":
          description: Request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendRequest"
  /friends/requests/{requestId}/accept:
    post:
      operationId: acceptFriendRequest
      tags: [friends]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /friends/requests/{requestId}/decline:
    post:
      operationId: declineFriendRequest
      tags: [friends]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Request declined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /friends/{friendId}:
    delete:
      operationId: removeFriend
      tags: [friends]
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Friend removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /blocks:
    get:
      operationId: getBlockedPlayers
      tags: [friends]
      responses:
        "200":
          description: Blocked players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Friend"
    post:
      operationId: blockPlayer
      tags: [friends]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsernameInput"
      responses:
        "201":
          description: Player blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /blocks/{playerId}:
    delete:
      operationId: unblockPlayer
      tags: [friends]
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Player unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
components:
  schemas:
...
    AccountCredentials:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, maxLength: 16, pattern: "^[A-Za-z0-9_]+$" }
        password: { type: string, minLength: 10, maxLength: 128 }
    Player:
      type: object
      required: [id, username, presence, characterSkin, isGuest]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        presence: { type: string, enum: [online, offline, in-game] }
        characterSkin: { type: string }
        isGuest: { type: boolean }
    AuthResult:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/Player" }
    ActionResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
    UsernameInput:
      type: object
      required: [username]
      properties:
        username: { type: string, minLength: 3, maxLength: 16 }
    Friend:
      type: object
      required: [userId, username, presence]
      properties:
        userId: { type: string, format: uuid }
        username: { type: string }
        presence: { type: string, enum: [online, offline, in-game] }
    FriendRequest:
      type: object
      required: [id, userId, username, createdAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        username: { type: string }
        createdAt: { type: string, format: date-time }
...