# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from './player';

export interface AuthResult {
  user: Player;
}