# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerPresence } from './playerPresence';

export interface Player {
  id: string;
  username: string;
  presence: PlayerPresence;
  characterSkin: string;
  isGuest: boolean;
}